.history/requirements_20230417232656.txt
absl-py==1.4.0
astunparse==1.6.3
beautifulsoup4==4.12.2
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
confection==0.0.4
cymem==2.0.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
Flask==2.2.3
flatbuffers==23.3.3
gast==0.4.0
geojson==2.5.0
google-auth==2.17.3
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
idna==3.4
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
langcodes==3.3.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
MouseInfo==0.1.3
murmurhash==1.0.9
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pathy==0.10.1
Pillow==9.5.0
preshed==3.0.8
protobuf==4.22.3
psutil==5.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
PyAutoGUI==0.9.53
pydantic==1.10.7
PyGetWindow==0.0.9
pyjokes==0.6.0
PyMsgBox==1.0.9
pyowm==3.3.0
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.0
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pyttsx3==2.90
pytweening==1.0.4
pytz==2023.3
pywhatkit==5.4
pywin32==306
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SpeechRecognition==3.10.0
srsly==2.4.6
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
thinc==8.1.9
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
wasabi==1.1.1
Werkzeug==2.2.3
wikipedia==1.4.0
wrapt==1.14.1
